# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movena replacement core contracts",
    "version": "2.0.0-alpha.1",
    "description": "Typed initial recovery slice. Other compatibility endpoints remain inventoried in baseline.json."
  },
  "paths": {
...
    "/api/v2/patient/data-rights-requests": {
      "get": {
        "operationId": "getDataRightsRequests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataRightsRecord"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      },
      "post": {
        "operationId": "createDataRightsRequest",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRightsRecord"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataRightsCreate"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/analysis-jobs": {
      "get": {
        "operationId": "getActiveAnalysisJobs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisJobResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataRightsCreate": {
        "properties": {
          "request_type": {
            "enum": [
              "export",
              "correction",
              "deletion"
            ],
            "title": "Request Type",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          }
        },
        "required": [
          "request_type"
        ],
        "title": "DataRightsCreate",
        "type": "object"
      },
      "DataRightsRecord": {
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "request_type": {
            "enum": [
              "export",
              "correction",
              "deletion"
            ],
            "title": "Request Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Completed At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Reviewed At"
          },
          "retention_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Retention Until"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Account Id"
          },
          "account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Account Email"
          },
          "account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Account Name"
          }
        },
        "required": [
          "request_id",
          "request_type",
          "status",
          "created_at"
        ],
        "title": "DataRightsRecord",
        "type": "object"
      },
...
      "AnalysisJobResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "exercise_id": {
            "title": "Exercise Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "progress": {
            "maximum": 100,
            "minimum": 0,
            "title": "Progress",
            "type": "integer"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Message"
          },
          "http_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Http Status"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Result"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Completed At"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Engine Version"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Model Version"
          }
        },
        "required": [
          "job_id",
          "exercise_id",
          "status",
          "stage",
          "progress",
          "attempts",
          "max_attempts",
          "created_at"
        ],
        "title": "AnalysisJobResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}